+++ app/src/androidTest/java/ch/epfl/sdp/appart/CameraUITest.java
@@ -62,10 +62,12 @@ public class CameraUITest {
                     public Matcher<View> getConstraints() {
                         return ViewMatchers.isEnabled(); // no constraints, they are checked above
                     }
+
                     @Override
                     public String getDescription() {
                         return "click gallery button";
                     }
+
                     @Override
                     public void perform(UiController uiController, View view) {
                         view.performClick();
