+++ app/src/androidTest/java/ch/epfl/sdp/appart/CreateUserAccountUITest.java
@@ -1,18 +1,11 @@
 package ch.epfl.sdp.appart;
 
 
-import android.view.View;
-import android.view.ViewGroup;
-import android.view.ViewParent;
-import androidx.test.espresso.ViewInteraction;
 import androidx.test.espresso.action.ViewActions;
 import androidx.test.espresso.intent.Intents;
 import androidx.test.espresso.matcher.ViewMatchers;
 import androidx.test.ext.junit.rules.ActivityScenarioRule;
 
-import org.hamcrest.Description;
-import org.hamcrest.Matcher;
-import org.hamcrest.TypeSafeMatcher;
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Rule;
@@ -35,18 +28,11 @@ import dagger.hilt.android.testing.UninstallModules;
 
 import static androidx.test.espresso.Espresso.onView;
 import static androidx.test.espresso.action.ViewActions.click;
-import static androidx.test.espresso.action.ViewActions.closeSoftKeyboard;
-import static androidx.test.espresso.action.ViewActions.replaceText;
 import static androidx.test.espresso.action.ViewActions.typeText;
-import static androidx.test.espresso.assertion.ViewAssertions.matches;
 import static androidx.test.espresso.intent.Intents.intended;
 import static androidx.test.espresso.intent.matcher.IntentMatchers.hasComponent;
-import static androidx.test.espresso.matcher.ViewMatchers.isDisplayed;
 import static androidx.test.espresso.matcher.ViewMatchers.withId;
-import static androidx.test.espresso.matcher.ViewMatchers.withParent;
-import static androidx.test.espresso.matcher.ViewMatchers.withText;
 import static org.hamcrest.MatcherAssert.assertThat;
-import static org.hamcrest.Matchers.allOf;
 import static org.hamcrest.Matchers.is;
 import static org.junit.Assert.assertNotNull;
 import static org.junit.Assert.assertNull;
@@ -91,6 +77,12 @@ public class CreateUserAccountUITest {
     }
 
     @Test
+    public void backToLoginTest() {
+        onView(withId(R.id.create_account_login_CreateUser_button)).perform(click());
+        intended(hasComponent(LoginActivity.class.getName()));
+    }
+
+    @Test
     public void successfulCreateAccountTest() throws ExecutionException, InterruptedException {
         String email = "test@testappart.ch";
         String password = "password";
@@ -113,78 +105,6 @@ public class CreateUserAccountUITest {
 
         loginService.deleteUser().get();
         assertNull(loginService.getCurrentUser());
-    }
-
-    @Test
-    public void createUserUITest() {
-        ViewInteraction appCompatEditText = onView(
-            allOf(withId(R.id.create_account_email_CreateUser_editText),
-                childAtPosition(
-                    childAtPosition(
-                        withId(android.R.id.content),
-                        0),
-                    2),
-                isDisplayed()));
-        appCompatEditText.perform(replaceText("test@epfl.ch"), closeSoftKeyboard());
-
-        ViewInteraction appCompatEditText2 = onView(
-            allOf(withId(R.id.create_account_password_CreateUser_editText),
-                childAtPosition(
-                    childAtPosition(
-                        withId(android.R.id.content),
-                        0),
-                    1),
-                isDisplayed()));
-        appCompatEditText2.perform(replaceText("1234567890"), closeSoftKeyboard());
-
-        ViewInteraction editText = onView(
-            allOf(withId(R.id.create_account_email_CreateUser_editText), withText("test@epfl.ch"),
-                withParent(withParent(withId(android.R.id.content))),
-                isDisplayed()));
-        editText.check(matches(withText("test@epfl.ch")));
-
-        ViewInteraction editText2 = onView(
-            allOf(withId(R.id.create_account_email_CreateUser_editText), withText("test@epfl.ch"),
-                withParent(withParent(withId(android.R.id.content))),
-                isDisplayed()));
-        editText2.check(matches(isDisplayed()));
-
-        ViewInteraction editText3 = onView(
-            allOf(withId(R.id.create_account_password_CreateUser_editText), withText("••••••••••"),
-                withParent(withParent(withId(android.R.id.content))),
-                isDisplayed()));
-        editText3.check(matches(withText("••••••••••")));
-
-        ViewInteraction editText4 = onView(
-            allOf(withId(R.id.create_account_password_CreateUser_editText), withText("••••••••••"),
-                withParent(withParent(withId(android.R.id.content))),
-                isDisplayed()));
-        editText4.check(matches(isDisplayed()));
-
-        ViewInteraction button = onView(
-            allOf(withId(R.id.create_account_CreateUser_button), withText("Create Account"),
-                withParent(withParent(withId(android.R.id.content))),
-                isDisplayed()));
-        button.check(matches(isDisplayed()));
-    }
-
-    private static Matcher<View> childAtPosition(
-        final Matcher<View> parentMatcher, final int position) {
-
-        return new TypeSafeMatcher<View>() {
-            @Override
-            public void describeTo(Description description) {
-                description.appendText("Child at position " + position + " in parent ");
-                parentMatcher.describeTo(description);
-            }
-
-            @Override
-            public boolean matchesSafely(View view) {
-                ViewParent parent = view.getParent();
-                return parent instanceof ViewGroup && parentMatcher.matches(parent)
-                    && view.equals(((ViewGroup) parent).getChildAt(position));
-            }
-        };
     }
 
     @After
