+++ app/src/main/java/ch/epfl/sdp/appart/ResetActivity.java
@@ -51,4 +51,15 @@ public class ResetActivity extends AppCompatActivity {
             startActivity(intent);
         });
     }
+
+    /**
+     * Method called when the Log In button is pushed
+     * Takes the user back to the login screen
+     *
+     * @param view
+     */
+    public void backToLogin(View view) {
+        Intent intent = new Intent(this, LoginActivity.class);
+        startActivity(intent);
+    }
 }
