+++ app/src/main/res/layout/activity_user_profile.xml
@@ -11,21 +11,30 @@
     tools:context=".UserProfileActivity">
 
     <LinearLayout
-      android:layout_width="0dp"
-      android:layout_height="match_parent"
-      android:paddingTop="70dp"
-      android:gravity="fill"
-      android:orientation="vertical"
-      app:layout_constraintBottom_toBottomOf="parent"
-      app:layout_constraintEnd_toEndOf="@+id/guideline26"
-      app:layout_constraintStart_toStartOf="@+id/guideline25"
-      app:layout_constraintTop_toTopOf="parent"
-      >
+        android:layout_width="0dp"
+        android:layout_height="match_parent"
+        android:paddingTop="70dp"
+        android:paddingBottom="10dp"
+        android:gravity="fill"
+        android:orientation="vertical"
+        app:layout_constraintBottom_toBottomOf="parent"
+        app:layout_constraintEnd_toEndOf="@+id/guideline26"
+        app:layout_constraintStart_toStartOf="@+id/guideline25"
+        app:layout_constraintTop_toTopOf="parent"
+        android:clipToPadding="false"
+        android:padding="2dp"
+        >
+
 
         <LinearLayout
-          android:layout_width="match_parent"
-          android:layout_height="wrap_content"
-          android:orientation="horizontal">
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:layout_gravity="fill"
+            android:clipToPadding="false"
+            android:orientation="horizontal"
+            android:paddingStart="16dp"
+            android:paddingTop="44dp"
+            >
 
             <LinearLayout
                 android:layout_width="wrap_content"
@@ -151,22 +160,25 @@
         </LinearLayout>
 
         <Space
-          android:layout_width="match_parent"
-          android:layout_height="0dp"
-          android:layout_weight="0.75" />
+            android:layout_width="match_parent"
+            android:layout_height="0dp"
+            android:layout_weight="1" />
 
         <LinearLayout
-          android:layout_width="match_parent"
-          android:layout_height="wrap_content"
-          android:gravity="bottom|center"
-          android:orientation="vertical">
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:layout_gravity="bottom"
+            android:foregroundGravity="bottom"
+            android:gravity="bottom"
+            android:orientation="vertical"
+            android:paddingBottom="16dp">
 
             <LinearLayout
-              android:id="@+id/attributes_UserProfile_layout"
-              android:layout_width="match_parent"
-              android:layout_height="wrap_content"
-              android:gravity="bottom|center"
-              android:orientation="vertical">
+                android:id="@+id/attributes_UserProfile_layout"
+                android:layout_width="match_parent"
+                android:layout_height="wrap_content"
+                android:orientation="vertical"
+                android:paddingBottom="53dp">
 
                 <LinearLayout
                     android:id="@+id/name_UserProfile_layout"
@@ -334,50 +346,44 @@
             </LinearLayout>
 
             <androidx.constraintlayout.widget.ConstraintLayout
-              android:layout_width="match_parent"
-              android:layout_height="wrap_content"
-              android:layout_gravity="bottom|center"
-              android:paddingBottom="20dp"
-              android:paddingTop="20dp">
+                android:layout_width="match_parent"
+                android:layout_height="wrap_content"
+                android:layout_gravity="bottom|center"
+                android:paddingBottom="20dp"
+                android:paddingTop="20dp"
+                android:clipToPadding="false"
+                android:padding="3dp">
 
                 <Button
-                  android:id="@+id/doneEditing_UserProfile_button"
-                  android:layout_width="match_parent"
-                  android:layout_height="wrap_content"
-                  android:layout_marginTop="40dp"
-                  android:layout_marginBottom="40dp"
-                  android:layout_marginStart="@dimen/margin_button"
-                  android:layout_marginEnd="@dimen/margin_button"
-                  android:text="@string/doneButtonText"
-                  android:background="@drawable/style_primary"
-                  android:textColor="#FFFFFF"
-                  android:textSize="16sp"
-                  android:textAllCaps="false"
-                  android:onClick="doneEditing"
-                  app:layout_constraintBottom_toBottomOf="parent"
-                  app:layout_constraintTop_toTopOf="parent"
-
-                  />
+                    android:id="@+id/doneEditing_UserProfile_button"
+                    android:layout_width="match_parent"
+                    android:layout_height="wrap_content"
+                    android:background="@drawable/style_primary"
+                    android:textColor="#FFFFFF"
+                    android:gravity="center"
+                    android:onClick="doneEditing"
+                    android:text="@string/doneButtonText"
+                    android:textAlignment="center"
+                    android:visibility="gone"
+                    android:textAllCaps="false"
+                    android:textSize="16sp"
+                    app:layout_constraintBottom_toBottomOf="parent" />
 
                 <Button
-                  android:id="@+id/editProfile_UserProfile_button"
-                  android:layout_width="match_parent"
-                  android:layout_height="wrap_content"
-                  android:layout_marginTop="40dp"
-                  android:layout_marginBottom="40dp"
-                  android:layout_marginStart="@dimen/margin_button"
-                  android:layout_marginEnd="@dimen/margin_button"
-                  android:text="@string/editProfileButtonText"
-                  android:paddingStart="40dp"
-                  android:paddingEnd="16dp"
-                  android:background="@drawable/style_secondary"
-                  android:drawableEnd="@drawable/ic_edit_user"
-                  android:textColor="@color/primaryStart"
-                  android:textSize="16sp"
-                  android:onClick="editProfile"
-                  android:textAllCaps="false"
-                  app:layout_constraintBottom_toBottomOf="parent"
-                  app:layout_constraintTop_toTopOf="parent"/>
+                    android:id="@+id/editProfile_UserProfile_button"
+                    android:layout_width="match_parent"
+                    android:layout_height="wrap_content"
+                    android:paddingStart="40dp"
+                    android:paddingEnd="16dp"
+                    android:background="@drawable/style_secondary"
+                    android:drawableEnd="@drawable/ic_edit_user"
+                    android:gravity="center"
+                    android:onClick="editProfile"
+                    android:text="@string/editProfileButtonText"
+                    android:textAllCaps="false"
+                    android:textColor="@color/primaryStart"
+                    app:layout_constraintBottom_toBottomOf="parent"
+                    android:textSize="16sp"/>
 
             </androidx.constraintlayout.widget.ConstraintLayout>
         </LinearLayout>
