+++ app/src/main/res/values/strings.xml
@@ -62,13 +62,7 @@
     <string name="changeProfilePicture"> CHANGE </string>
     <string name="updateUserFailedInDB">  Information update failed, try again </string>
     <string name="updateUserImageFailedInDB">  Image update failed, try again </string>
-    <string-array name="genderList">
-        <item> </item>
-        <item>Female</item>
-        <item>Male</item>
-        <item>Other</item>
-    </string-array>
-    <string name="contactAdUser">Contact</string>
+    <string name="contactAdUser"> CONTACT ANNOUNCER </string>
     <string name="localSaveFailed_User">Failed to save updates locally!</string>
     <string name="loadFailed_User">Failed to load user information</string>
     <string name="cameraAdButtonDesc">Add Picture</string>
@@ -155,6 +149,12 @@
     <string name="add_to_favorites">Add to Favorites</string>
     <string name="favorites">Favorites</string>
     <string name="userAds">Your Ads</string>
+    <string-array name="genderList">
+        <item> </item>
+        <item>Female</item>
+        <item>Male</item>
+        <item>Other</item>
+    </string-array>
 
     <!-- FullScreenImageActivity stuff -->
     <string-array name="periodList">
